Disable the check for path self-intersection
It is very expensive, unlikely (but not impossible) to happen, and not terrible if it does happen. Either we get a knot in the rail, which might lead to some jams, or the build will fail.

This provides a significant boost to long range pathfinding speed (my torture test map goes from >14 to <2 years to find a path).