Add lake detector!
When trying to find a spot for a station, look at the four edges of the rectangle defined by the start (crossing exit) and destination (station entrance). Checks the tiles along each edge for long runs of water; if more than the max, consider that edge "blocked by water". However, this is only problematic if *two parallel edges* are blocked. In that case, it will declare it an unviable spot for a station, and look elsewhere.